requirements.txt
streamlit
python-docx
requests
openai
google-generativeai
pydantic
pymupdf  # for PDF processing
python-docx